refactor code of redirect of authentication page according to user is signIn/SignUp
