+++ netflix-app/src/Component/Body.js
@@ -6,13 +6,11 @@ import {
   createBrowserRouter,
   useNavigate,
 } from "react-router-dom";
-import { getAuth, onAuthStateChanged } from "firebase/auth";
-import { auth } from "../Utils/firebase_config";
-import { useDispatch } from "react-redux";
-import { addUser, removeUser } from "../Utils/Slice/userSlice";
+
+
 
 const Body = () => {
-  const dispatch = useDispatch();
+  
 
   const appRouter = createBrowserRouter([
     {
@@ -25,30 +23,7 @@ const Body = () => {
     },
   ]);
 
-  useEffect(() => {
-    onAuthStateChanged(auth, (user) => {
-      if (user) {
-        // User is signed in, see docs for a list of available properties
-        // https://firebase.google.com/docs/reference/js/auth.user
-
-        const { uid, displayName, email, photoURL } = user;
-        dispatch(
-          addUser({
-            uid: uid,
-            displayName: displayName,
-            email: email,
-            photoURL: photoURL,
-          })
-        );
-
-        // ...
-      } else {
-        // User is signed out
-        // ...
-        dispatch(removeUser());
-      }
-    });
-  }, []);
+ 
   return (
     <>
       <RouterProvider router={appRouter} />
