+++ netflix-app/src/Utils/constant.js
@@ -1,0 +1,9 @@
+export const Netflix_logo="https://cdn.cookielaw.org/logos/dd6b162f-1a32-456a-9cfe-897231c7763c/4345ea78-053c-46d2-b11e-09adaef973dc/Netflix_Logo_PMS.png"
+export const LoginPage_backgrd="https://assets.nflxext.com/ffe/siteui/vlv3/f85718e8-fc6d-4954-bca0-f5eaf78e0842/ea44b42b-ba19-4f35-ad27-45090e34a897/IN-en-20230918-popsignuptwoweeks-perspective_alpha_website_large.jpg"
+export const API_options= {
+    method: 'GET',
+    headers: {
+      accept: 'application/json',
+      Authorization: 'Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiJmNWZlZmUyNzVhYWU3Y2U0MWRjOGVhOWRiZTg2ZTA3OSIsInN1YiI6IjYzZTBlM2MwODc0MWM0MDA5OWI1NWRiYiIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.6WGCQf3e-YU-701l-n9HMez-eZC8adbEob3t7ABkE7k'
+    }
+  };
